requirements.txt
# Web Framework
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0

# Database & ORM
sqlalchemy>=2.0.0,<3.0.0

# Environment Variables
python-dotenv>=1.0.0,<2.0.0

# Cryptography (Token Encryption)
cryptography>=42.0.0,<43.0.0

# Secure signing for cookies
itsdangerous>=2.1,<3.0

# HTTP Requests
requests>=2.31.0,<3.0.0

# Type hinting
typing-extensions>=4.0.0
